backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic[email]
python-dotenv
google-generativeai
pdfplumber
python-multipart